Squashed an awful lot of warnings, very few left now. Added iterator retrieval methods for entities, cameras, lights, billboardsets to SceneManager Node::removeChild and SceneNode::removeAndDestroyAllChildren now calls needUpdate to update bounding box Added cubemapsJS.zip back to all resources.cfg now it is ok clearScene now reinitialises nodes for skies Added setNormaliseNormals to RenderSystem and added option to Renderable and Entity, to allow people to bypass lighting problems with scaling if they want

+++ OgreMain/include/OgreListChanger.h
@@ -45,7 +45,7 @@ namespace Ogre {
 		virtual ResourceListConstIterator getConstIterator() = 0;
 		virtual ResourceListConstIterator getConstEndIterator() = 0;
 		virtual  Resource* popFront() = 0;
-		virtual  int getListSize() = 0;
+		virtual  size_t getListSize() = 0;
     };
 
 } // namespace
