+++ RenderSystems/Direct3D9/src/OgreD3D9HardwareVertexBuffer.cpp
@@ -36,7 +36,7 @@ namespace Ogre {
     {
         // Create the vertex buffer
         HRESULT hr = pDev->CreateVertexBuffer(
-            mSizeInBytes, 
+            static_cast<UINT>(mSizeInBytes), 
             D3D9Mappings::get(usage), 
             0, // No FVF here, thankyou
 			useSystemMemory? D3DPOOL_SYSTEMMEM : D3DPOOL_DEFAULT, 
@@ -71,8 +71,8 @@ namespace Ogre {
 			lockOpts= D3D9Mappings::get(options);
 		} 
         HRESULT hr = mlpD3DBuffer->Lock(
-            offset, 
-            length, 
+            static_cast<UINT>(offset), 
+            static_cast<UINT>(length), 
             &pBuf,
             lockOpts);
